[REFACTOR]: api 테스트 패키지 위치 이동
