+++ src/test/java/project/domain/user/api/UserApiTest.java
@@ -1,4 +1,4 @@
-package project.api;
+package project.domain.user.api;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
@@ -9,7 +9,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import org.junit.jupiter.api.*;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
-import project.common.IntegrationTest;
+import project.common.helper.IntegrationTest;
 
 public class UserApiTest extends IntegrationTest {
